Convert posts library to async TypeScript
